requirements.txt
python-dotenv
gspread
google-auth
aiogram==3.*